requirements.txt
Django==5.0
django-cors-headers==4.3.0
djangorestframework==3.14.0
Pillow==10.4.0
python-telegram-bot==21.0
python-dotenv==1.0.0
qrcode==7.4.2

